chore: add namespace to App/index.d.ts
