+++ src/components/Card/CellCard.tsx
@@ -3,11 +3,10 @@ import styled from 'styled-components'
 import { Link } from 'react-router-dom'
 import AppContext from '../../contexts/App'
 import CopyGreenIcon from '../../assets/copy_green.png'
-import { InputOutput, Script, Data, Wrapper } from '../../types/App/index'
 import { startEndEllipsis, hexToUtf8 } from '../../utils/string'
 import { shannonToCkb, copyElementValue } from '../../utils/util'
-import { CellType, fetchScript, fetchCellData } from '../../service/fetcher'
-import { Response } from '../../types/App/Response'
+import { CellType, fetchScript, fetchCellData } from '../../service/http/fetcher'
+import { Response, Wrapper } from '../../service/http/Response'
 import { localeNumberString } from '../../utils/number'
 import i18n from '../../utils/i18n'
 
@@ -175,10 +174,10 @@ const CardLabelItem = ({
   )
 }
 
-const Capacity = ({ cell }: { cell: InputOutput }) => (
+const Capacity = ({ cell }: { cell: State.InputOutput }) => (
   <CardLabelItem name="Capacity" value={localeNumberString(shannonToCkb(cell.capacity))} />
 )
-const CellAddressCapacityItem = ({ type, cell }: { type: CellType; cell: InputOutput }) => {
+const CellAddressCapacityItem = ({ type, cell }: { type: CellType; cell: State.InputOutput }) => {
   const name = type === CellType.Input ? 'Input' : 'Output'
 
   if (cell.from_cellbase) {
@@ -254,7 +253,7 @@ const getCellState = (state: any, cellState: CellState) => {
   return cellState === state.cellState ? CellState.NONE : cellState
 }
 
-const CellScriptItem = ({ cellType, cell }: { cellType: CellType; cell: InputOutput }) => {
+const CellScriptItem = ({ cellType, cell }: { cellType: CellType; cell: State.InputOutput }) => {
   const appContext = useContext(AppContext)
   const [state, dispatch] = useReducer(reducer, initialState)
 
@@ -285,22 +284,22 @@ const CellScriptItem = ({ cellType, cell }: { cellType: CellType; cell: InputOut
     switch (getCellState(state, cellState)) {
       case CellState.LOCK:
         fetchScript(cellType, 'lock_scripts', `${cell.id}`).then(response => {
-          const { data } = response as Response<Wrapper<Script>>
+          const { data } = response as Response<Wrapper<State.Script>>
           handleCellState(cellState)
           showScriptContent(data ? data.attributes : initScriptContent.lock)
         })
         break
       case CellState.TYPE:
         fetchScript(cellType, 'type_scripts', `${cell.id}`).then(response => {
-          const { data } = response as Response<Wrapper<Script>>
+          const { data } = response as Response<Wrapper<State.Script>>
           handleCellState(cellState)
           showScriptContent(data ? data.attributes : initScriptContent.type)
         })
         break
       case CellState.DATA:
-        fetchCellData(cellType, `${cell.id}`).then((data: Data) => {
+        fetchCellData(cellType, `${cell.id}`).then((data: State.Data) => {
           if (data.data.length < MAX_DATA_SIZE) {
-            const dataValue: Data = data
+            const dataValue: State.Data = data
             if (data && cell.isGenesisOutput) {
               dataValue.data = hexToUtf8(data.data.substr(2))
             }
@@ -356,7 +355,7 @@ const CellScriptItem = ({ cellType, cell }: { cellType: CellType; cell: InputOut
   )
 }
 
-const CellCard = ({ type, cell }: { type: CellType; cell: InputOutput }) => {
+const CellCard = ({ type, cell }: { type: CellType; cell: State.InputOutput }) => {
   return (
     <CardPanel>
       <CellAddressCapacityItem type={type} cell={cell} />
