+++ src/components/Transaction/TransactionCell/index.tsx
@@ -1,7 +1,6 @@
 import React, { useState } from 'react'
 import { Link } from 'react-router-dom'
 import HelpIcon from '../../../assets/qa_help.png'
-import { InputOutput } from '../../../types/App/index'
 import i18n from '../../../utils/i18n'
 import { localeNumberString } from '../../../utils/number'
 import { startEndEllipsis } from '../../../utils/string'
@@ -24,7 +23,7 @@ const Cellbase = ({ targetBlockNumber }: { targetBlockNumber?: number }) => {
   }
   return (
     <CellbasePanel>
-      <div className="cellbase__content">Cellbase for Block</div>
+      <div className="cellbase__content">Cellbase for State.Block</div>
       <Link to={`/block/${targetBlockNumber}`}>{localeNumberString(targetBlockNumber)}</Link>
       <div
         className="cellbase__help"
@@ -52,7 +51,7 @@ const Cellbase = ({ targetBlockNumber }: { targetBlockNumber?: number }) => {
   )
 }
 
-const TransactionCell = ({ cell, address }: { cell: InputOutput; address?: string }) => {
+const TransactionCell = ({ cell, address }: { cell: State.InputOutput; address?: string }) => {
   if (cell.from_cellbase) {
     return <Cellbase targetBlockNumber={cell.target_block_number} />
   }
