+++ src/pages/BlockList/index.tsx
@@ -21,9 +21,8 @@ import TransactionIcon from '../../assets/transactions.png'
 import BlockRewardIcon from '../../assets/block_reward_white.png'
 import MinerIcon from '../../assets/miner.png'
 import TimestampIcon from '../../assets/timestamp.png'
-import { fetchBlockList } from '../../service/fetcher'
-import { Block, Wrapper } from '../../types/App/index'
-import { Response } from '../../types/App/Response'
+import { fetchBlockList } from '../../service/http/fetcher'
+import { Response, Wrapper } from '../../service/http/Response'
 import { shannonToCkb } from '../../utils/util'
 import { parsePageNumber } from '../../utils/string'
 import { CachedKeys } from '../../utils/const'
@@ -61,7 +60,7 @@ const reducer = (state: any, action: any) => {
 
 const getBlocks = (page: number, size: number, dispatch: any) => {
   fetchBlockList(page, size).then(response => {
-    const { data, meta } = response as Response<Wrapper<Block>[]>
+    const { data, meta } = response as Response<Wrapper<State.Block>[]>
     if (meta) {
       dispatch({
         type: Actions.total,
@@ -83,7 +82,7 @@ const getBlocks = (page: number, size: number, dispatch: any) => {
 }
 
 const initialState = {
-  blocks: [] as Wrapper<Block>[],
+  blocks: [] as Wrapper<State.Block>[],
   total: 1,
 }
 
@@ -100,7 +99,7 @@ export default (props: React.PropsWithoutRef<RouteComponentProps>) => {
   const [state, dispatch] = useReducer(reducer, initialState)
 
   useEffect(() => {
-    const cachedBlocks = fetchCachedData<Wrapper<Block>[]>(CachedKeys.BlockList)
+    const cachedBlocks = fetchCachedData<Wrapper<State.Block>[]>(CachedKeys.BlockList)
     if (cachedBlocks) {
       dispatch({
         type: Actions.blocks,
@@ -146,7 +145,7 @@ export default (props: React.PropsWithoutRef<RouteComponentProps>) => {
               <TableTitleItem image={TimestampIcon} title={t('home.time')} />
             </TableTitleRow>
             {state.blocks &&
-              state.blocks.map((data: Wrapper<Block>) => {
+              state.blocks.map((data: Wrapper<State.Block>) => {
                 return (
                   data && (
                     <TableContentRow key={data.attributes.block_hash}>
