+++ src/pages/Home/index.tsx
@@ -18,9 +18,8 @@ import MinerIcon from '../../assets/miner.png'
 import TimestampIcon from '../../assets/timestamp.png'
 import MoreLeftIcon from '../../assets/more_left.png'
 import MoreRightIcon from '../../assets/more_right.png'
-import { fetchBlocks, fetchStatistics } from '../../service/fetcher'
-import { Block, Statistics, Wrapper } from '../../types/App/index'
-import { Response } from '../../types/App/Response'
+import { fetchBlocks, fetchStatistics } from '../../service/http/fetcher'
+import { Response, Wrapper } from '../../service/http/Response'
 import { shannonToCkb } from '../../utils/util'
 import { parseTime, parseSimpleDate } from '../../utils/date'
 import { BLOCK_POLLING_TIME, CachedKeys } from '../../utils/const'
@@ -44,14 +43,14 @@ const BlockchainItem = ({ name, value, tip }: { name: string; value: string; tip
 
 const getLatestBlocks = (setBlocksWrappers: any) => {
   fetchBlocks().then(response => {
-    const { data } = response as Response<Wrapper<Block>[]>
+    const { data } = response as Response<Wrapper<State.Block>[]>
     setBlocksWrappers(data)
   })
 }
 
 const getStatistics = (setStatistics: any) => {
   fetchStatistics().then(response => {
-    const { data } = response as Response<Wrapper<Statistics>>
+    const { data } = response as Response<Wrapper<State.Statistics>>
     setStatistics(data.attributes)
   })
 }
@@ -62,7 +61,7 @@ interface BlockchainData {
   tip: string
 }
 
-const initStatistics: Statistics = {
+const initStatistics: State.Statistics = {
   tip_block_number: '0',
   average_block_time: '0',
   current_epoch_difficulty: 0,
@@ -74,17 +73,17 @@ const parseHashRate = (hashRate: string | undefined) => {
 }
 
 export default () => {
-  const initBlockWrappers: Wrapper<Block>[] = []
+  const initBlockWrappers: Wrapper<State.Block>[] = []
   const [blocksWrappers, setBlocksWrappers] = useState(initBlockWrappers)
   const [statistics, setStatistics] = useState(initStatistics)
   const [t] = useTranslation()
 
   useEffect(() => {
-    const cachedBlocks = fetchCachedData<Wrapper<Block>[]>(CachedKeys.Blocks)
+    const cachedBlocks = fetchCachedData<Wrapper<State.Block>[]>(CachedKeys.Blocks)
     if (cachedBlocks) {
       setBlocksWrappers(cachedBlocks)
     }
-    const cachedStatistics = fetchCachedData<Statistics>(CachedKeys.Statistics)
+    const cachedStatistics = fetchCachedData<State.Statistics>(CachedKeys.Statistics)
     if (cachedStatistics) {
       setStatistics(cachedStatistics)
     }
