+++ src/pages/Transaction/Script.tsx
@@ -2,9 +2,8 @@ import React, { useContext, useReducer } from 'react'
 import { Link } from 'react-router-dom'
 import AppContext from '../../contexts/App'
 import CopyGreenIcon from '../../assets/copy_green.png'
-import { Response } from '../../types/App/Response'
-import { Script, InputOutput, Data, Wrapper } from '../../types/App/index'
-import { CellType, fetchScript, fetchCellData } from '../../service/fetcher'
+import { Response, Wrapper } from '../../service/http/Response'
+import { CellType, fetchScript, fetchCellData } from '../../service/http/fetcher'
 import { shannonToCkb, copyElementValue } from '../../utils/util'
 import { hexToUtf8, parseLongAddressHash } from '../../utils/string'
 import { localeNumberString } from '../../utils/number'
@@ -49,7 +48,7 @@ const reducer = (state: any, action: any) => {
   }
 }
 
-const ScriptTypeItems = ['Lock Script', 'Type Script', 'Data']
+const ScriptTypeItems = ['Lock State.Script', 'Type State.Script', 'State.Data']
 const cellStateWithItem = (item: string) => {
   if (item === ScriptTypeItems[0]) {
     return CellState.LOCK
@@ -67,7 +66,7 @@ const getCellState = (state: any, item: string) => {
   return cellState === state.cellState ? CellState.NONE : cellState
 }
 
-const AddressHashComponent = ({ cell }: { cell: InputOutput }) => {
+const AddressHashComponent = ({ cell }: { cell: State.InputOutput }) => {
   return cell.address_hash ? (
     <Link to={`/address/${cell.address_hash}`}>
       <code>{parseLongAddressHash(cell.address_hash)}</code>
@@ -77,7 +76,7 @@ const AddressHashComponent = ({ cell }: { cell: InputOutput }) => {
   )
 }
 
-const ScriptComponent = ({ cellType, cell }: { cellType: CellType; cell: InputOutput }) => {
+const ScriptComponent = ({ cellType, cell }: { cellType: CellType; cell: State.InputOutput }) => {
   const appContext = useContext(AppContext)
   const [state, dispatch] = useReducer(reducer, initialState)
 
@@ -107,20 +106,20 @@ const ScriptComponent = ({ cellType, cell }: { cellType: CellType; cell: InputOu
     switch (getCellState(state, item)) {
       case CellState.LOCK:
         fetchScript(cellType, 'lock_scripts', `${cell.id}`).then(response => {
-          const { data } = response as Response<Wrapper<Script>>
+          const { data } = response as Response<Wrapper<State.Script>>
           handleCellState(item)
           showScriptContent(data ? data.attributes : initScriptContent.lock)
         })
         break
       case CellState.TYPE:
         fetchScript(cellType, 'type_scripts', `${cell.id}`).then(response => {
-          const { data } = response as Response<Wrapper<Script>>
+          const { data } = response as Response<Wrapper<State.Script>>
           handleCellState(item)
           showScriptContent(data ? data.attributes : initScriptContent.type)
         })
         break
       case CellState.DATA:
-        fetchCellData(cellType, `${cell.id}`).then((data: Data) => {
+        fetchCellData(cellType, `${cell.id}`).then((data: State.Data) => {
           const dataValue = data
           if (data && cell.isGenesisOutput) {
             dataValue.data = hexToUtf8(data.data.substr(2))
