+++ src/pages/Transaction/index.tsx
@@ -21,9 +21,9 @@ import TransactionIcon from '../../assets/transaction_fee.png'
 import CopyIcon from '../../assets/copy.png'
 import StatusIcon from '../../assets/transcation_status.png'
 import { parseSimpleDate } from '../../utils/date'
-import { Response } from '../../types/App/Response'
-import { Transaction, InputOutput, Statistics, Wrapper } from '../../types/App/index'
-import { CellType, fetchTransactionByHash, fetchTipBlockNumber } from '../../service/fetcher'
+import { Response, Wrapper } from '../../service/http/Response'
+
+import { CellType, fetchTransactionByHash, fetchTipBlockNumber } from '../../service/http/fetcher'
 import { copyElementValue, formatConfirmation, shannonToCkb } from '../../utils/util'
 import CellCard from '../../components/Card/CellCard'
 import ScriptComponent from './Script'
@@ -75,7 +75,7 @@ const InputOutputTableTitle = ({ transactionType, isCellbase }: { transactionTyp
   )
 }
 
-const initTransaction: Transaction = {
+const initTransaction: State.Transaction = {
   transaction_hash: '',
   block_number: 0,
   block_timestamp: 0,
@@ -90,8 +90,8 @@ const initTransaction: Transaction = {
 const getTransaction = (hash: string, setTransaction: any, replace: any) => {
   fetchTransactionByHash(hash)
     .then(response => {
-      const { data } = response as Response<Wrapper<Transaction>>
-      const transactionValue = data.attributes as Transaction
+      const { data } = response as Response<Wrapper<State.Transaction>>
+      const transactionValue = data.attributes as State.Transaction
       if (transactionValue.display_outputs && transactionValue.display_outputs.length > 0) {
         transactionValue.display_outputs[0].isGenesisOutput = transactionValue.block_number === 0
       }
@@ -104,7 +104,7 @@ const getTransaction = (hash: string, setTransaction: any, replace: any) => {
 
 const getTipBlockNumber = (setTipBlockNumber: any) => {
   fetchTipBlockNumber().then(response => {
-    const { data } = response as Response<Wrapper<Statistics>>
+    const { data } = response as Response<Wrapper<State.Statistics>>
     if (data) {
       setTipBlockNumber(parseInt(data.attributes.tip_block_number, 10))
     }
@@ -181,13 +181,13 @@ export default (props: React.PropsWithoutRef<RouteComponentProps<{ hash: string 
           <div>
             {transaction &&
               transaction.display_inputs &&
-              transaction.display_inputs.map((input: InputOutput, index: number) => {
+              transaction.display_inputs.map((input: State.InputOutput, index: number) => {
                 const key = index
                 return <CellCard type={CellType.Input} cell={input} key={key} />
               })}
             {transaction &&
               transaction.display_outputs &&
-              transaction.display_outputs.map((output: InputOutput, index: number) => {
+              transaction.display_outputs.map((output: State.InputOutput, index: number) => {
                 const key = index
                 return <CellCard type={CellType.Output} cell={output} key={key} />
               })}
@@ -209,7 +209,7 @@ export default (props: React.PropsWithoutRef<RouteComponentProps<{ hash: string 
                 <tbody>
                   {transaction &&
                     transaction.display_inputs &&
-                    transaction.display_inputs.map((input: InputOutput) => {
+                    transaction.display_inputs.map((input: State.InputOutput) => {
                       return input && <ScriptComponent cellType={CellType.Input} key={input.id} cell={input} />
                     })}
                 </tbody>
@@ -222,7 +222,7 @@ export default (props: React.PropsWithoutRef<RouteComponentProps<{ hash: string 
                 <tbody>
                   {transaction &&
                     transaction.display_outputs &&
-                    transaction.display_outputs.map((output: InputOutput) => {
+                    transaction.display_outputs.map((output: State.InputOutput) => {
                       return output && <ScriptComponent cellType={CellType.Output} key={output.id} cell={output} />
                     })}
                 </tbody>
