+++ src/service/http/fetcher.ts
@@ -1,5 +1,5 @@
 import axios, { AxiosResponse } from 'axios'
-import CONFIG from '../config'
+import CONFIG from '../../config'
 
 const baseURL = `${CONFIG.API_URL}/api/v1/`
