+++ src/utils/util.ts
@@ -1,4 +1,3 @@
-import { Transaction, InputOutput } from '../types/App/index'
 import { MAX_CONFIRMATION } from './const'
 import i18n from './i18n'
 import { parseNumber } from './number'
@@ -20,17 +19,17 @@ export const shannonToCkb = (value: number) => {
   return value / 10 ** 8
 }
 
-const handleCellCapacity = (cells: InputOutput[], address?: string) => {
+const handleCellCapacity = (cells: State.InputOutput[], address?: string) => {
   if (!cells || cells.length === 0) return 0
   return cells
-    .filter((cell: InputOutput) => cell.address_hash === address)
-    .map((cell: InputOutput) => parseNumber(cell.capacity))
+    .filter((cell: State.InputOutput) => cell.address_hash === address)
+    .map((cell: State.InputOutput) => parseNumber(cell.capacity))
     .reduce((previous: number, current: number) => {
       return previous + current
     }, 0)
 }
 
-export const handleCapacityChange = (transaction: Transaction, address?: string) => {
+export const handleCapacityChange = (transaction: State.Transaction, address?: string) => {
   if (!transaction) return 0
   return (
     handleCellCapacity(transaction.display_outputs, address) - handleCellCapacity(transaction.display_inputs, address)
